+++ changelog.md
@@ -54,3 +54,8 @@
 * added several new icons
 * added contact section in Landing.tsx
 * added contributors.json to hold all contact people
+
+## 22.06.2024
+* added global context to hold the current section scroll index
+* implemented scroll logic in Navbar component
+* added indexes to each link in [searchOptions.json]